Producer consumer benchmark

+++ relaxed_concurrent_fifo/concurrent_fifo.h
@@ -30,7 +30,7 @@ private:
 	}
 
 public:
-	concurrent_fifo([[maybe_unused]] size_t thread_count, size_t capacity) : capacity(capacity) {
+	concurrent_fifo([[maybe_unused]] int thread_count, size_t capacity) : capacity(capacity) {
 		if (!is_po2(capacity)) {
 			throw std::runtime_error("Please only use capacities that are a power of two");
 		}
